Relocate Project table model into the model layer
Standardize on LiteLLM_ProjectTable, now defined in litellm/models/project.py
(referencing the relocated Budget and ObjectPermission models) and re-exported
from litellm.proxy._types. Repoint ProjectRepository; drop the hand-rolled
Project domain model. Teach the test MockTable to synthesize primary keys on
create so repositories returning PK-required canonical types are covered.
