+++ litellm/models/project.py
@@ -1,32 +1,40 @@
 """
-Project domain model.
+Project table model.
+
+Canonical definition for ``litellm_projecttable``. Re-exported from
+``litellm.proxy._types`` for backwards compatibility.
 """
 
-from typing import Any, Dict, List, Optional
+from datetime import datetime
+from typing import List, Optional
 
-from pydantic import Field
+from litellm.models.budget import LiteLLM_BudgetTable
+from litellm.models.object_permission import LiteLLM_ObjectPermissionTable
+from litellm.types.llms.base import LiteLLMPydanticObjectBase
 
-from litellm.models.base import DomainModel
 
+class LiteLLM_ProjectTable(LiteLLMPydanticObjectBase):
+    """Database model representation for project"""
 
-class Project(DomainModel):
-    """Domain model for projects."""
-
-    project_id: Optional[str] = None
+    project_id: str
     project_alias: Optional[str] = None
     description: Optional[str] = None
     team_id: Optional[str] = None
     budget_id: Optional[str] = None
-    metadata: Dict[str, Any] = Field(default_factory=dict)
-    models: List[str] = Field(default_factory=list)
+    metadata: Optional[dict] = None
+    models: List[str] = []
     spend: float = 0.0
-    model_spend: Dict[str, float] = Field(default_factory=dict)
-    model_rpm_limit: Dict[str, int] = Field(default_factory=dict)
-    model_tpm_limit: Dict[str, int] = Field(default_factory=dict)
+    model_spend: Optional[dict] = None
+    model_rpm_limit: Optional[dict] = None
+    model_tpm_limit: Optional[dict] = None
     blocked: bool = False
     object_permission_id: Optional[str] = None
     created_by: Optional[str] = None
     updated_by: Optional[str] = None
+    created_at: Optional[datetime] = None
+    updated_at: Optional[datetime] = None
+    litellm_budget_table: Optional[LiteLLM_BudgetTable] = None
+    object_permission: Optional[LiteLLM_ObjectPermissionTable] = None
 
     @property
     def is_blocked(self) -> bool:
