+++ litellm/repositories/project_repository.py
@@ -4,11 +4,11 @@ Project repository for database operations on LiteLLM_ProjectTable.
 
 from typing import Any, Dict, List, Optional, Type
 
-from litellm.models.project import Project
+from litellm.models.project import LiteLLM_ProjectTable
 from litellm.repositories.base_repository import BaseRepository
 
 
-class ProjectRepository(BaseRepository[Project]):
+class ProjectRepository(BaseRepository[LiteLLM_ProjectTable]):
     """Repository for project database operations."""
 
     @property
@@ -16,22 +16,22 @@ class ProjectRepository(BaseRepository[Project]):
         return self.prisma_client.db.litellm_projecttable
 
     @property
-    def model_class(self) -> Type[Project]:
-        return Project
+    def model_class(self) -> Type[LiteLLM_ProjectTable]:
+        return LiteLLM_ProjectTable
 
     async def find_by_id(
         self, project_id: str, id_field: str = "project_id"
-    ) -> Optional[Project]:
+    ) -> Optional[LiteLLM_ProjectTable]:
         return await super().find_by_id(project_id, id_field)
 
-    async def find_by_alias(self, project_alias: str) -> Optional[Project]:
+    async def find_by_alias(self, project_alias: str) -> Optional[LiteLLM_ProjectTable]:
         """Find a project by alias."""
         records = await self.table.find_many(where={"project_alias": project_alias})
         if records:
             return self._to_model(records[0])
         return None
 
-    async def find_by_team_id(self, team_id: str) -> List[Project]:
+    async def find_by_team_id(self, team_id: str) -> List[LiteLLM_ProjectTable]:
         """Find all projects belonging to a team."""
         records = await self.table.find_many(where={"team_id": team_id})
         return self._to_model_list(records)
@@ -49,7 +49,7 @@ class ProjectRepository(BaseRepository[Project]):
         model_rpm_limit: Optional[Dict[str, int]] = None,
         model_tpm_limit: Optional[Dict[str, int]] = None,
         object_permission_id: Optional[str] = None,
-    ) -> Project:
+    ) -> LiteLLM_ProjectTable:
         """Create a new project."""
         data: Dict[str, Any] = {
             "created_by": created_by,
@@ -92,7 +92,7 @@ class ProjectRepository(BaseRepository[Project]):
         model_tpm_limit: Optional[Dict[str, int]] = None,
         blocked: Optional[bool] = None,
         object_permission_id: Optional[str] = None,
-    ) -> Optional[Project]:
+    ) -> Optional[LiteLLM_ProjectTable]:
         """Update a project."""
         data: Dict[str, Any] = {"updated_by": updated_by}
         if project_alias is not None:
@@ -118,10 +118,12 @@ class ProjectRepository(BaseRepository[Project]):
 
         return await self.update(project_id, data, id_field="project_id")
 
-    async def delete_project(self, project_id: str) -> Optional[Project]:
+    async def delete_project(self, project_id: str) -> Optional[LiteLLM_ProjectTable]:
         """Delete a project."""
         return await self.delete(project_id, id_field="project_id")
 
-    async def update_spend(self, project_id: str, spend: float) -> Optional[Project]:
+    async def update_spend(
+        self, project_id: str, spend: float
+    ) -> Optional[LiteLLM_ProjectTable]:
         """Update project spend."""
         return await self.update(project_id, {"spend": spend}, id_field="project_id")
